config/requirements.txt
# ============================================================================
# AIS Vessel Tracker - Pinned Dependencies
# ============================================================================
# All versions pinned for reproducibility and production stability.
# Generated: 2025-11-20
# Python: 3.13+
# ============================================================================

# Core Web Framework
# ----------------------------------------------------------------------------
Flask==3.1.2                    # Web framework for vessel tracking interface
Flask-SocketIO==5.5.1           # Real-time WebSocket support for live updates
Werkzeug==3.1.3                 # WSGI utility library (Flask dependency)
Jinja2==3.1.6                   # Template engine for Flask
MarkupSafe==3.0.3               # Safe string handling for Jinja2
click==8.3.0                    # CLI framework (Flask dependency)
itsdangerous==2.2.0             # Secure data signing (Flask dependency)

# Real-time Communication
# ----------------------------------------------------------------------------
python-socketio==5.12.0         # Socket.IO server implementation
python-engineio==4.10.1         # Engine.IO server (socketio dependency)
websocket-client==1.8.0         # WebSocket client for AISStream API
simple-websocket==1.1.0         # WebSocket wrapper (socketio dependency)
wsproto==1.2.0                  # WebSocket protocol implementation

# HTTP Requests & Web Scraping
# ----------------------------------------------------------------------------
requests==2.32.5                # HTTP library for API calls
urllib3==2.5.0                  # HTTP client (requests dependency)
certifi==2025.10.5              # SSL certificates
charset-normalizer==3.4.3       # Character encoding detection
idna==3.10                      # Internationalized domain names
beautifulsoup4==4.14.2          # HTML/XML parsing for company lookup
soupsieve==2.8                  # CSS selector library (bs4 dependency)

# Data Processing & Analysis
# ----------------------------------------------------------------------------
pandas==2.3.3                   # Data manipulation for emissions analysis
numpy==2.3.3                    # Numerical computing (pandas dependency)
openpyxl==3.1.5                 # Excel file handling for EU MRV data
et-xmlfile==2.0.0               # XML utilities (openpyxl dependency)
python-dateutil==2.9.0.post0    # Date parsing utilities
pytz==2025.2                    # Timezone support
tzdata==2025.2                  # Timezone database

# AI & Machine Learning (Optional)
# ----------------------------------------------------------------------------
google-generativeai==0.8.5      # Google Gemini API for company profiling
google-ai-generativelanguage==0.6.15  # Gemini API types
google-api-core==2.28.1         # Google API client core
google-api-python-client==2.186.0     # Google API client
google-auth==2.42.1             # Google authentication
google-auth-httplib2==0.2.1     # Google auth HTTP support
googleapis-common-protos==1.71.0      # Google API protocol buffers
proto-plus==1.26.1              # Protocol buffer wrappers
protobuf==5.29.5                # Protocol buffers
grpcio==1.76.0                  # gRPC framework
grpcio-status==1.71.2           # gRPC status codes
cachetools==6.2.1               # Caching utilities (google-auth dependency)
pyasn1==0.6.1                   # ASN.1 types (google-auth dependency)
pyasn1-modules==0.4.2           # ASN.1 protocol modules
rsa==4.9.1                      # RSA encryption (google-auth dependency)

# Web Automation (Optional - for scraping)
# ----------------------------------------------------------------------------
playwright==1.55.0              # Browser automation for company scraping
greenlet==3.2.4                 # Concurrency (playwright dependency)

# Search & NLP (Optional)
# ----------------------------------------------------------------------------
googlesearch-python==1.3.0      # Google search API wrapper

# Testing & Development
# ----------------------------------------------------------------------------
pytest==9.0.1                   # Testing framework
pluggy==1.6.0                   # Plugin system (pytest dependency)
iniconfig==2.3.0                # INI file parser (pytest dependency)
packaging==25.0                 # Package version handling
flake8==7.3.0                   # Code linting for CI/CD
pycodestyle==2.14.0             # Style checker (flake8 dependency)
pyflakes==3.4.0                 # Static checker (flake8 dependency)
mccabe==0.7.0                   # Complexity checker (flake8 dependency)

# Utilities
# ----------------------------------------------------------------------------
colorama==0.4.6                 # Terminal colors (Windows compatibility)
tqdm==4.67.1                    # Progress bars for long operations
python-dotenv==1.0.1            # Environment variable management

# ============================================================================
# Installation Instructions
# ============================================================================
#
# Fresh install:
#   python -m venv venv
#   source venv/bin/activate  # Windows: venv\Scripts\activate
#   pip install -r config/requirements.txt
#
# Upgrade existing:
#   pip install -r config/requirements.txt --upgrade
#
# Export current environment:
#   pip freeze > config/requirements-freeze.txt
#
# ============================================================================
